Changed path to configuration file from ./test/texttx.conf to ./testtx.conf.
Changed some numeric literals to explicit float types; do not know if it changes any result, but it is more correct.

Changed "omega" vector name to "w" becasue that's what's in the document and the "omega" name was confusing; by convention in DSP, "omega" usually represents an angular frequency, not a signal name.



git-svn-id: http://voip.null.ro/svn/bladeRF@17 6180954b-3207-0510-9d3d-99f8faf60350